Fix cache system not working properly
Clean up codes.
