+++ src/config/binding.h
@@ -3,6 +3,7 @@
 
 #include <toml.hpp>
 
+#include "../handler/settings.h"
 #include "crontask.h"
 #include "proxygroup.h"
 #include "regmatch.h"
@@ -171,29 +172,6 @@ namespace toml
 
 namespace INIBinding
 {
-
-    void parseGroupTimes(const std::string &src, int *interval, int *tolerance, int *timeout)
-    {
-        std::vector<int*> ptrs;
-        ptrs.push_back(interval);
-        ptrs.push_back(timeout);
-        ptrs.push_back(tolerance);
-        string_size bpos = 0, epos = src.find(",");
-        for(int *x : ptrs)
-        {
-            if(x != NULL)
-                *x = to_int(src.substr(bpos, epos - bpos), 0);
-            if(epos != src.npos)
-            {
-                bpos = epos + 1;
-                epos = src.find(",", bpos);
-            }
-            else
-                return;
-        }
-        return;
-    }
-
     template<class T> struct from
     {};
