+++ src/generator/config/ruleconvert.h
@@ -18,7 +18,7 @@ enum ruleset_type
     RULESET_CLASH_CLASSICAL
 };
 
-struct ruleset_content
+struct RulesetContent
 {
     std::string rule_group;
     std::string rule_path;
@@ -29,8 +29,8 @@ struct ruleset_content
 };
 
 std::string convertRuleset(const std::string &content, int type);
-void rulesetToClash(YAML::Node &base_rule, std::vector<ruleset_content> &ruleset_content_array, bool overwrite_original_rules, bool new_field_name);
-std::string rulesetToClashStr(YAML::Node &base_rule, std::vector<ruleset_content> &ruleset_content_array, bool overwrite_original_rules, bool new_field_name);
-void rulesetToSurge(INIReader &base_rule, std::vector<ruleset_content> &ruleset_content_array, int surge_ver, bool overwrite_original_rules, std::string remote_path_prefix);
+void rulesetToClash(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_content_array, bool overwrite_original_rules, bool new_field_name);
+std::string rulesetToClashStr(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_content_array, bool overwrite_original_rules, bool new_field_name);
+void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_content_array, int surge_ver, bool overwrite_original_rules, std::string remote_path_prefix);
 
 #endif // RULECONVERT_H_INCLUDED
