+++ src/handler/interfaces.h
@@ -9,7 +9,9 @@
 #include "../generator/config/subexport.h"
 #include "../server/webserver.h"
 
-void refreshRulesets(RulesetConfigs &ruleset_list, std::vector<ruleset_content> &rca);
+std::string parseProxy(const std::string &source);
+
+void refreshRulesets(RulesetConfigs &ruleset_list, std::vector<RulesetContent> &rca);
 void readConf();
 int simpleGenerator();
 std::string convertRuleset(const std::string &content, int type);
