+++ src/main.cpp
@@ -9,6 +9,7 @@
 #include "config/ruleset.h"
 #include "handler/interfaces.h"
 #include "handler/webget.h"
+#include "handler/settings.h"
 #include "script/cron.h"
 #include "server/socket.h"
 #include "server/webserver.h"
@@ -22,13 +23,6 @@
 #include "version.h"
 
 //#include "vfs.h"
-
-extern std::string gPrefPath, gAccessToken, gListenAddress, gGenerateProfiles, gManagedConfigPrefix;
-extern bool gAPIMode, gGeneratorMode, gCFWChildProcess, gUpdateRulesetOnRequest;
-extern int gListenPort, gMaxConcurThreads, gMaxPendingConns;
-extern RulesetConfigs gCustomRulesets;
-extern std::vector<ruleset_content> gRulesetContent;
-extern bool gEnableCron;
 
 WebServer webServer;
 
@@ -70,22 +64,22 @@ void chkArg(int argc, char *argv[])
     {
         if(strcmp(argv[i], "-cfw") == 0)
         {
-            gCFWChildProcess = true;
-            gUpdateRulesetOnRequest = true;
+            global.CFWChildProcess = true;
+            global.updateRulesetOnRequest = true;
         }
         else if(strcmp(argv[i], "-f") == 0 || strcmp(argv[i], "--file") == 0)
         {
             if(i < argc - 1)
-                gPrefPath.assign(argv[++i]);
+                global.prefPath.assign(argv[++i]);
         }
         else if(strcmp(argv[i], "-g") == 0 || strcmp(argv[i], "--gen") == 0)
         {
-            gGeneratorMode = true;
+            global.generatorMode = true;
         }
         else if(strcmp(argv[i], "--artifact") == 0)
         {
             if(i < argc - 1)
-                gGenerateProfiles.assign(argv[++i]);
+                global.generateProfiles.assign(argv[++i]);
         }
         else if(strcmp(argv[i], "-l") == 0 || strcmp(argv[i], "--log") == 0)
         {
@@ -115,7 +109,7 @@ void signal_handler(int sig)
 
 void cron_tick_caller()
 {
-    if(gEnableCron)
+    if(global.enableCron)
         cron_tick();
 }
 
@@ -126,26 +120,26 @@ int main(int argc, char *argv[])
     setcd(prgpath); //first switch to program directory
 #endif // _DEBUG
     if(fileExist("pref.toml"))
-        gPrefPath = "pref.toml";
+        global.prefPath = "pref.toml";
     else if(fileExist("pref.yml"))
-        gPrefPath = "pref.yml";
+        global.prefPath = "pref.yml";
     else if(!fileExist("pref.ini"))
     {
         if(fileExist("pref.example.toml"))
         {
             fileCopy("pref.example.toml", "pref.toml");
-            gPrefPath = "pref.toml";
+            global.prefPath = "pref.toml";
         }
-        if(fileExist("pref.example.yml"))
+        else if(fileExist("pref.example.yml"))
         {
             fileCopy("pref.example.yml", "pref.yml");
-            gPrefPath = "pref.yml";
+            global.prefPath = "pref.yml";
         }
         else if(fileExist("pref.example.ini"))
             fileCopy("pref.example.ini", "pref.ini");
     }
     chkArg(argc, argv);
-    setcd(gPrefPath); //then switch to pref directory
+    setcd(global.prefPath); //then switch to pref directory
     writeLog(0, "SubConverter " VERSION " starting up..", LOG_LEVEL_INFO);
 #ifdef _WIN32
     WSADATA wsaData;
@@ -169,17 +163,18 @@ int main(int argc, char *argv[])
 
     SetConsoleTitle("SubConverter " VERSION);
     readConf();
-    if(!gUpdateRulesetOnRequest)
-        refreshRulesets(gCustomRulesets, gRulesetContent);
+    //vfs::vfs_read("vfs.ini");
+    if(!global.updateRulesetOnRequest)
+        refreshRulesets(global.customRulesets, global.rulesetsContent);
 
     std::string env_api_mode = getEnv("API_MODE"), env_managed_prefix = getEnv("MANAGED_PREFIX"), env_token = getEnv("API_TOKEN");
-    gAPIMode = tribool().parse(toLower(env_api_mode)).get(gAPIMode);
+    global.APIMode = tribool().parse(toLower(env_api_mode)).get(global.APIMode);
     if(env_managed_prefix.size())
-        gManagedConfigPrefix = env_managed_prefix;
+        global.managedConfigPrefix = env_managed_prefix;
     if(env_token.size())
-        gAccessToken = env_token;
+        global.accessToken = env_token;
 
-    if(gGeneratorMode)
+    if(global.generatorMode)
         return simpleGenerator();
 
     /*
@@ -196,42 +191,42 @@ int main(int argc, char *argv[])
 
     webServer.append_response("GET", "/refreshrules", "text/plain", [](RESPONSE_CALLBACK_ARGS) -> std::string
     {
-        if(gAccessToken.size())
+        if(global.accessToken.size())
         {
             std::string token = getUrlArg(request.argument, "token");
-            if(token != gAccessToken)
+            if(token != global.accessToken)
             {
                 response.status_code = 403;
                 return "Forbidden\n";
             }
         }
-        refreshRulesets(gCustomRulesets, gRulesetContent);
+        refreshRulesets(global.customRulesets, global.rulesetsContent);
         return "done\n";
     });
 
     webServer.append_response("GET", "/readconf", "text/plain", [](RESPONSE_CALLBACK_ARGS) -> std::string
     {
-        if(gAccessToken.size())
+        if(global.accessToken.size())
         {
             std::string token = getUrlArg(request.argument, "token");
-            if(token != gAccessToken)
+            if(token != global.accessToken)
             {
                 response.status_code = 403;
                 return "Forbidden\n";
             }
         }
         readConf();
-        if(!gUpdateRulesetOnRequest)
-            refreshRulesets(gCustomRulesets, gRulesetContent);
+        if(!global.updateRulesetOnRequest)
+            refreshRulesets(global.customRulesets, global.rulesetsContent);
         return "done\n";
     });
 
     webServer.append_response("POST", "/updateconf", "text/plain", [](RESPONSE_CALLBACK_ARGS) -> std::string
     {
-        if(gAccessToken.size())
+        if(global.accessToken.size())
         {
             std::string token = getUrlArg(request.argument, "token");
-            if(token != gAccessToken)
+            if(token != global.accessToken)
             {
                 response.status_code = 403;
                 return "Forbidden\n";
@@ -239,9 +234,9 @@ int main(int argc, char *argv[])
         }
         std::string type = getUrlArg(request.argument, "type");
         if(type == "form")
-            fileWrite(gPrefPath, getFormData(request.postdata), true);
+            fileWrite(global.prefPath, getFormData(request.postdata), true);
         else if(type == "direct")
-            fileWrite(gPrefPath, request.postdata, true);
+            fileWrite(global.prefPath, request.postdata, true);
         else
         {
             response.status_code = 501;
@@ -249,14 +244,14 @@ int main(int argc, char *argv[])
         }
 
         readConf();
-        if(!gUpdateRulesetOnRequest)
-            refreshRulesets(gCustomRulesets, gRulesetContent);
+        if(!global.updateRulesetOnRequest)
+            refreshRulesets(global.customRulesets, global.rulesetsContent);
         return "done\n";
     });
 
     webServer.append_response("GET", "/flushcache", "text/plain", [](RESPONSE_CALLBACK_ARGS) -> std::string
     {
-        if(getUrlArg(request.argument, "token") != gAccessToken)
+        if(getUrlArg(request.argument, "token") != global.accessToken)
         {
             response.status_code = 403;
             return "Forbidden";
@@ -283,7 +278,7 @@ int main(int argc, char *argv[])
 
     webServer.append_response("GET", "/convert", "text/plain;charset=utf-8", getConvertedRuleset);
 
-    if(!gAPIMode)
+    if(!global.APIMode)
     {
         webServer.append_response("GET", "/get", "text/plain;charset=utf-8", [](RESPONSE_CALLBACK_ARGS) -> std::string
         {
@@ -303,10 +298,10 @@ int main(int argc, char *argv[])
 
     std::string env_port = getEnv("PORT");
     if(env_port.size())
-        gListenPort = to_int(env_port, gListenPort);
-    listener_args args = {gListenAddress, gListenPort, gMaxPendingConns, gMaxConcurThreads, cron_tick_caller, 200};
+        global.listenPort = to_int(env_port, global.listenPort);
+    listener_args args = {global.listenAddress, global.listenPort, global.maxPendingConns, global.maxConcurThreads, cron_tick_caller, 200};
     //std::cout<<"Serving HTTP @ http://"<<listen_address<<":"<<listen_port<<std::endl;
-    writeLog(0, "Startup completed. Serving HTTP @ http://" + gListenAddress + ":" + std::to_string(gListenPort), LOG_LEVEL_INFO);
+    writeLog(0, "Startup completed. Serving HTTP @ http://" + global.listenAddress + ":" + std::to_string(global.listenPort), LOG_LEVEL_INFO);
     webServer.start_web_server_multi(&args);
 
 #ifdef _WIN32
